apps/worker/pyproject.toml
[project]
name = "indexa-worker"
version = "0.1.0"
description = "Indexa knowledge base vectorization worker"
requires-python = ">=3.12"
dependencies = [
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
  "qdrant-client>=1.12",
  "dashscope>=1.20",
  "pypdf>=5.1",
  "langchain-text-splitters>=0.3",
  "cuid2>=2.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/indexa_worker"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
